+++ src/core.c
@@ -481,7 +481,7 @@ SEXP rnng_dial(SEXP socket, SEXP url, SEXP autostart, SEXP error) {
 
   if (xc) {
     R_Free(dp);
-    if (error != R_NilValue) ERROR_OUT(xc);
+    if (LOGICAL(error)[0]) ERROR_OUT(xc);
     ERROR_RET(xc);
   }
 
@@ -531,7 +531,7 @@ SEXP rnng_listen(SEXP socket, SEXP url, SEXP autostart, SEXP error) {
   const int xc = start ? nng_listen(*sock, up, lp, 0) : nng_listener_create(lp, *sock, up);
   if (xc) {
     R_Free(lp);
-    if (error != R_NilValue) ERROR_OUT(xc);
+    if (LOGICAL(error)[0]) ERROR_OUT(xc);
     ERROR_RET(xc);
   }